requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
fair-esm>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Graph Processing
networkx>=3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
